CHannel > Category

+++ EGG9000.Bot/Automated/Coops/CoopStatusUpdater.cs
@@ -608,8 +608,8 @@ namespace EGG9000.Bot.Automated.Coops {
                     }
                 }
                 var currentUsers = await coopThread.ExtGetUsersAsync();
-                var threadCategory = await coopThread.GetCategoryAsync();
-                var overwrites = threadCategory.PermissionOverwrites.Where(p => p.Permissions.ViewChannel == PermValue.Allow);
+                var threadChannel = guild.TextChannels.First(c => c.Threads.Any(t => t.Id == coopThread.Id));
+                var overwrites = threadChannel.Category.PermissionOverwrites.Where(p => p.Permissions.ViewChannel == PermValue.Allow);
                 List<SocketGuildUser> usersContained = [];
                 foreach(var roleOverwrite in overwrites.Where(o => o.TargetType == PermissionTarget.Role)) {
                     usersContained.AddRange(guild.GetRole(roleOverwrite.TargetId).Members);
@@ -749,37 +749,39 @@ namespace EGG9000.Bot.Automated.Coops {
                 }
 
                 var pingsLeft = usersNeedingChannelPermissions.Select(id => $"<@{id}>").ToList();
-                if((await coopThread.GetPinnedMessagesAsync()).Where(m => m.Author.IsBot && m.Content != "\u17B5").LastOrDefault() is IUserMessage editPingsInto) {
-                    var currentContent = editPingsInto.Content;
-                    //Determine how many 22 character strings we can fit into content
-                    var pingsPerCycle = (1500 - currentContent.Length) / 22;
-                    while(pingsLeft.Count > 0) {
-                        // Remove pingsPerCycle entries from pingsLeft
-                        var pingsToRemove = pingsLeft.Take(pingsPerCycle).ToList();
-                        pingsLeft.RemoveRange(0, Math.Min(pingsPerCycle, pingsLeft.Count));
+                if(pingsLeft.Count > 0) {
+                    if((await coopThread.GetPinnedMessagesAsync()).Where(m => m.Author.IsBot && m.Content != "\u17B5").LastOrDefault() is IUserMessage editPingsInto) {
+                        var currentContent = editPingsInto.Content;
+                        //Determine how many 22 character strings we can fit into content
+                        var pingsPerCycle = (1500 - currentContent.Length) / 22;
+                        while(pingsLeft.Count > 0) {
+                            // Remove pingsPerCycle entries from pingsLeft
+                            var pingsToRemove = pingsLeft.Take(pingsPerCycle).ToList();
+                            pingsLeft.RemoveRange(0, Math.Min(pingsPerCycle, pingsLeft.Count));
 
-                        // Combine pingsToRemove into a message, append to current content
-                        var message = string.Join(" ", pingsToRemove);
+                            // Combine pingsToRemove into a message, append to current content
+                            var message = string.Join(" ", pingsToRemove);
 
-                        await editPingsInto.ModifyAsync(m => m.Content = currentContent + " " + message);
-                        await editPingsInto.ModifyAsync(m => m.Content = currentContent);
+                            await editPingsInto.ModifyAsync(m => m.Content = currentContent + " " + message);
+                            await editPingsInto.ModifyAsync(m => m.Content = currentContent);
+                        }
+                    } else {
+                        var response = await coopThread.SendMessageAsync("[Ping into]");
+                        //Determine how many 22 character strings we can fit into content
+                        var pingsPerCycle = 1500 / 22;
+                        while(pingsLeft.Count > 0) {
+                            // Remove pingsPerCycle entries from pingsLeft
+                            var pingsToRemove = pingsLeft.Take(pingsPerCycle).ToList();
+                            pingsLeft.RemoveRange(0, Math.Min(pingsPerCycle, pingsLeft.Count));
+
+                            // Combine pingsToRemove into a message, append to current content
+                            var message = string.Join(" ", pingsToRemove);
+
+                            await response.ModifyAsync(m => m.Content = message);
+                            await response.ModifyAsync(m => m.Content = "[Ping into]");
+                        }
+                        await response.DeleteAsync();
                     }
-                } else {
-                    var response = await coopThread.SendMessageAsync("[Ping into]");
-                    //Determine how many 22 character strings we can fit into content
-                    var pingsPerCycle = 1500 / 22;
-                    while(pingsLeft.Count > 0) {
-                        // Remove pingsPerCycle entries from pingsLeft
-                        var pingsToRemove = pingsLeft.Take(pingsPerCycle).ToList();
-                        pingsLeft.RemoveRange(0, Math.Min(pingsPerCycle, pingsLeft.Count));
-
-                        // Combine pingsToRemove into a message, append to current content
-                        var message = string.Join(" ", pingsToRemove);
-
-                        await response.ModifyAsync(m => m.Content = message);
-                        await response.ModifyAsync(m => m.Content = "[Ping into]");
-                    }
-                    await response.DeleteAsync();
                 }
                 var usersNow = await coopThread.ExtGetUsersAsync();
                 var usersAdded = usersNow.Where(un => !currentUsers.Contains(un)).Select(u => u.Id).ToList();
